straus scalar slice in new files
